requirements.txt
pillow>=12.0.0